+++ src/object/define.js
@@ -1,0 +1,5 @@
+import { defineInstance } from '../core/object';
+import * as methods from './methods';
+
+defineInstance('mapKeys', methods.mapKeys);
+defineInstance('mapValues', methods.mapValues);