+++ src/object/index.js
@@ -1,1 +1,3 @@
 export * from '../core/object';
+export * from './methods';
+import './define';