+++ src/object/methods.js
@@ -1,0 +1,2 @@
+export { default as mapKeys } from './mapKeys';
+export { default as mapValues } from './mapValues';